pyproject.toml
[project]
name = "googleaisearch2api"
version = "0.1.0"
description = "Expose Google AI search interactions as an OpenAI-compatible API using Patchright."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "patchright>=1.58.2",
    "pydantic-settings>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "ruff>=0.12.10",
]

[project.scripts]
googleaisearch2api = "googleaisearch2api.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/googleaisearch2api"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
